Moving network provider tools to services
