+++ android-agent/src/main/java/io/opentelemetry/android/internal/services/network/CurrentNetworkProvider.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.android.instrumentation.network;
+package io.opentelemetry.android.internal.services.network;
 
 import android.app.Application;
 import android.content.Context;
@@ -15,6 +15,9 @@ import android.os.Build;
 import android.util.Log;
 import androidx.annotation.NonNull;
 import io.opentelemetry.android.common.RumConstants;
+import io.opentelemetry.android.internal.services.network.data.CurrentNetwork;
+import io.opentelemetry.android.internal.services.network.data.NetworkState;
+import io.opentelemetry.android.internal.services.network.detector.NetworkDetector;
 import java.util.List;
 import java.util.concurrent.CopyOnWriteArrayList;
 import java.util.function.Supplier;
@@ -28,9 +31,9 @@ import java.util.function.Supplier;
  */
 public final class CurrentNetworkProvider {
 
-    static final CurrentNetwork NO_NETWORK =
+    public static final CurrentNetwork NO_NETWORK =
             CurrentNetwork.builder(NetworkState.NO_NETWORK_AVAILABLE).build();
-    static final CurrentNetwork UNKNOWN_NETWORK =
+    public static final CurrentNetwork UNKNOWN_NETWORK =
             CurrentNetwork.builder(NetworkState.TRANSPORT_UNKNOWN).build();
 
     /**
