+++ android-agent/src/main/java/io/opentelemetry/android/internal/services/network/NetworkChangeListener.java
@@ -3,7 +3,9 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.android.instrumentation.network;
+package io.opentelemetry.android.internal.services.network;
+
+import io.opentelemetry.android.internal.services.network.data.CurrentNetwork;
 
 interface NetworkChangeListener {
